+++ changelog.md
@@ -2,6 +2,8 @@
 
 ### 3.1.0 beta
 + Added: `sticky-footer`.
++ Added: `parallelogram`.
++ Added: `angled-edges`.
 + Improved: `grid` now works with `@media query` and can make several rows.
 + Improved: rewrite `container` with `css-calc`, now you can combine `container` with other structure tags:
 ```html